Money changed to double. Obviously not supported in Azure search :D

+++ Oogi/Oogi/Tokens/AzureSearch.cs
@@ -67,7 +67,7 @@ namespace Oogi.Tokens
         [IsFilterable]
         [IsSortable]
         [IsFacetable]
-        public decimal? Money { get; set; }
+        public double? Money { get; set; }
 
         [IsFilterable]
         [IsFacetable]
